+++ src/kits/network/compat/libbind/dst/md5_dgst.c
@@ -99,7 +99,7 @@ unsigned long len;
 	int sw,sc;
 	ULONG l;
 
-	if (len == 0) return;
+	if (len == 0U) return;
 
 	l=(c->Nl+(len<<3))&0xffffffffL;
 	/* 95-05-24 eay Fixed a bug with the overflow handling, thanks to
@@ -115,7 +115,7 @@ unsigned long len;
 		sw=c->num>>2;
 		sc=c->num&0x03;
 
-		if ((c->num+len) >= MD5_CBLOCK)
+		if ((c->num+len) >= (size_t)MD5_CBLOCK)
 			{
 			l= p[sw];
 			p_c2l(data,l,sc);
@@ -136,7 +136,7 @@ unsigned long len;
 			int ew,ec;
 
 			c->num+=(int)len;
-			if ((sc+len) < 4) /* ugly, add char's to a word */
+			if ((sc+len) < 4U) /* ugly, add char's to a word */
 				{
 				l= p[sw];
 				p_c2l_p(data,l,sc,len);
@@ -163,7 +163,7 @@ unsigned long len;
 	/* we now can process the input data in blocks of MD5_CBLOCK
 	 * chars and save the leftovers to c->data. */
 	p=c->data;
-	while (len >= MD5_CBLOCK)
+	while (len >= (size_t)MD5_CBLOCK)
 		{
 #if defined(L_ENDIAN) || defined(B_ENDIAN)
 		memcpy(p,data,MD5_CBLOCK);
